Agricraft to MyAgriculture Conversion
Supporting work for the additional of Mystical Agriculture and allowing
conversion of existing Agricraft resources to the new mod.

JEI Blacklist updated to hide anything made uncraftable, as well as
seeds
CraftTweaker scripts for converting seeds from Agricraft ones to MyAg
ones. Also removes crafting recipes for water channels, tanks, valves
and grates, as well as the sprinkler.

Notes that only seeds that have 3 stat values matching will be able to
be converted - so a 1/1/1 seed or a 5/5/5 seed would convert, be a 1/2/1
or a 5/8/7 would not.